backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.9.1",
    "python-jose>=3.4.0",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]
